requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
PyPDF2
python-docx
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
httplib2
